==> scATAC-seq/4-clustering/2-level_2/05-exclude_clusters_level_2.Rmd <==
---
author: "Paula Soler-Vila & Ramon Massoni-Badosa"
output: 
  BiocStyle::html_document:
    toc: true
    toc_float: true
    number_sections: true
params:
  cell_type: PC
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, out.width = "100%", fig.align='center', 
                      message=FALSE, warning = FALSE)
options(width = 1200)
title <- paste("Exclude doublets and poor-quality cells (level 2):", params$cell_type, sep = " ")
```

---
title: "`r title`"
---


# Introduction

The objective of this notebook is to exclude the clusters of doublets or poor-quality cells.


# Pre-processing

## Load packages

```{r}
library(Seurat)
library(tidyverse)
```


## Parameters

```{r}
# Paths
path_to_doublets <- here::here("scRNA-seq/3-clustering/2-level_2/tmp/doublets_multiome_df_all.rds")

path_to_dir <- here::here("scATAC-seq/results/R_objects/level_2/")
path_to_obj <- str_c(
  path_to_dir,
  params$cell_type,
  "/",
  params$cell_type,
  "_integrated_level_2.rds",
  sep = ""
)
path_to_save <- str_c(
  path_to_dir,
  params$cell_type,
  "/",
  params$cell_type,
  "_clustered_filtered_level_2.rds",
  sep = ""
)


# Functions
source(here::here("scRNA-seq/bin/utils.R"))


# Colors
color_palette <- c("black", "gray", "red", "yellow", "violet", "green4",
                   "blue", "chocolate1", "coral2", "blueviolet",
                   "brown1", "darkmagenta", "deepskyblue1", "dimgray",
                   "deeppink1", "green", "lightgray", "hotpink1",
                   "indianred4", "khaki", "mediumorchid2")

# Resolutions
resolutions <- c(
  NBC_MBC = 0.025,
  GCBC = 0.05,
  CD4_T = 0.025,
  Cytotoxic = 0.01,
  PC = 0.01
)
resolution <- resolutions[params$cell_type]
names(resolution) <- NULL


# Clusters to exclude
cluster_to_exclude_all <- list(
  NBC_MBC = c("1"),
  GCBC = c("1"),
  CD4_T = c(as.character(seq(1,3))),
  Cytotoxic = c(as.character(seq(1,3))),
  PC = c(as.character(seq(1,8)))
)
cluster_to_exclude <- cluster_to_exclude_all[[params$cell_type]]
```


## Load data

```{r}
# Seurat object
seurat <- readRDS(path_to_obj)
seurat
```

# Cluster

Here, we cluster with the resolution that allows us to separate our clusters of interest, which was decided by inspecting the results of previous notebooks.

```{r}
seurat <- FindClusters(seurat, resolution = resolution)
p <- DimPlot(seurat, cols = color_palette,label = T)
p
```


# Exclude doublets and poor-quality cells

## Exclude doublets directly from scRNAseq data using the Multiome approach

### Visualize before filtering

```{r fig.wide=TRUE}
multiome_doublets <- readRDS(path_to_doublets)
doublets_cells <- colnames(seurat)[which(colnames(seurat) %in% multiome_doublets$barcode)]
length(doublets_cells)

DimPlot(
  seurat, reduction = "umap",
  cols.highlight = "darkred", cols= "grey",
  cells.highlight= doublets_cells,
  pt.size = 0.1
)
```

### Visualize after filtering

```{r fig.wide=TRUE}
selected_cells_doublets <- colnames(seurat)[which(!colnames(seurat) %in% doublets_cells)]
seurat <- subset(seurat, cells = selected_cells_doublets)
seurat

DimPlot(
  seurat, reduction = "umap",
  pt.size = 0.1
)
```

## Exclude problematic clusters

```{r}
print("Clusters to exclude")
cluster_to_exclude

selected_cells <- colnames(seurat)[!(seurat$seurat_clusters %in% cluster_to_exclude)]
length(selected_cells)

seurat <- subset(seurat, cells = selected_cells)
seurat
```


### Visualize after filtering

```{r}
p2 <- DimPlot(seurat, cols = color_palette, label = T)
p2
```


# Save

```{r}
saveRDS(seurat, path_to_save)
```


# Session Information

```{r}
sessionInfo()
```
